Squashed commit of the following:
commit 55caaa021ab6453a34f88d30256f6f4f2c0bcd58
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 23:05:51 2017 +0800

    ES查询有问题，待修复

commit d849246a9d42bf4fda8600504b98ce1243bde5d8
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 22:06:39 2017 +0800

    es or查询修改

commit fe578650086521d5a6733d285c37e539642e050b
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 20:22:33 2017 +0800

    es是否有手机号判断

commit 0a45294f42cd5c002cc684a683015cda114423fe
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 20:20:05 2017 +0800

    ES是否招聘信息的条件加入

commit dc40e41bbe258f53bec8f90fb708454d1c654701
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 18:33:44 2017 +0800

    推送成功历史

commit ab711b674ea921ebc8ffbf22901051c8e0110075
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 17:29:17 2017 +0800

    推送成功历史，加上部门和人员

commit c71095ab929ea2239582903a1570c299a02808f0
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 16:41:45 2017 +0800

    推送成功历史

commit 5177d0e0fa94344c8d24f649a1605a9e254d4d69
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 16:13:58 2017 +0800

    推送成功历史记录联系时长统计

commit e4374e413ae15e926da6fef05f738e91c07401cb
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 14:14:41 2017 +0800

    jobName要带上，查询个人信息

commit 4f0bc0173c9ae86f98049d607935c6cd3b936a7f
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 11:04:45 2017 +0800

    1

commit 0b36ddb080768241c6d0bfdd70796df7185ba940
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 10:30:13 2017 +0800

    修改dockerfile

commit 317273a8d689eae8d145a4110afb275c0da2efe4
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 10:28:55 2017 +0800

    获取个人信息

commit a28e5e2389099efe4ffc413759eb1d8cc74ce0fe
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 10:19:48 2017 +0800

    获取个人信息

commit 010f9cbb0772452589b6a3fd2ced1fec9eb2aef8
Author: wuwf <272551766@qq.com>
Date:   Tue Nov 14 10:16:15 2017 +0800

    获取个人信息

commit f76cd540afdea70acaefd110ca765f1e8bb695d1
Author: wuwf <272551766@qq.com>
Date:   Mon Nov 13 19:23:08 2017 +0800

    查询es时带上mobile 为空的条件

commit 3ea9835fb662e976dbd5ffa0e525ffc12635875b
Author: wuwf <272551766@qq.com>
Date:   Sun Nov 12 12:57:04 2017 +0800

    查询es时带上state=0的条件

commit 5d4d51cc7655a1eef8fa0bef2a181e25042496e2
Author: wuwf <272551766@qq.com>
Date:   Sat Nov 11 13:19:18 2017 +0800

    body toStirng

commit b5a49e64329da9001badb5a603aea3ec470e7dc5
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 18:15:27 2017 +0800

    body toStirng

commit 2554a7254c09a1d1e5df6d15666210e253c70127
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 18:08:21 2017 +0800

    全文检索

commit ac14a53d2f21cede552be8fc561c14ee3d9ad820
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 17:48:00 2017 +0800

    全文检索

commit 6d4685c5d2ea26d3e9c2763fd820071a5ed4634c
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 16:34:06 2017 +0800

    啥也没干

commit 8a75d60619499a1de4add12c4fe2d258b8cf222f
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 16:22:40 2017 +0800

    es查的信息，orderby走默认

commit 3387e52e60505854e0c2bf24ec66dec8bc5a39c3
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 15:59:02 2017 +0800

    丰富contact的信息，从es读取以部门

commit da463166f8098ad34de331673138f8ec382fb1cc
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 12:46:02 2017 +0800

    1

commit 4b707781885fb5e23c4f1b8229f4e1ce5ed432a3
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 12:14:33 2017 +0800

    es存的state都不为0，查不到，先注释掉

commit ca5196ae8bd50a7953d45fecefe188544dcc63c3
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 11:50:24 2017 +0800

    es功能可用了

commit 8dd9576b0a23aa7fcab917d71c2654af48546d28
Author: wuwf <272551766@qq.com>
Date:   Fri Nov 10 11:36:52 2017 +0800

    es不存在时创建index

# Conflicts:
#	logback.log

+++ src/main/java/com/mindata/ecserver/main/controller/DepartmentController.java
@@ -33,7 +33,7 @@ public class DepartmentController {
     @GetMapping("")
     @RequiresRoles(value = {Constant.ROLE_MANAGER, Constant.ROLE_ADMIN, Constant.ROLE_LEADER}, logical = Logical.OR)
     public BaseData find(@PageableDefault(sort = {"sort", "id"}, direction = Sort.Direction.DESC) Pageable pageable,
-                         String name, Integer companyId) {
+                         String name, Long companyId) {
         return ResultGenerator.genSuccessResult(departmentService.findByName(companyId, name, pageable));
     }
