+++ src/main/java/com/mindata/ecserver/main/repository/secondary/PtMenuRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * @author wuweifeng wrote on 2017/10/26.
  */
-public interface PtMenuRepository extends JpaRepository<PtMenu, Integer> {
+public interface PtMenuRepository extends JpaRepository<PtMenu, Long> {
 }
